abstract function to flash/draw attention to searchbar for reusability
