+++ webapp/src/components/Slider/NewEvent.js
@@ -1,6 +1,7 @@
 import { useState, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import FlashSearchbar from '../../utils/FlashSearchbar';
 import { HardSetList } from '../../store/reel';
 import { UnfixMap, Focus } from '../../store/map';
 import { CreateEvent, ShowLast } from '../../store/homeSlider';
@@ -18,7 +19,7 @@ export default function NewEventModal () {
   const [title, updateTitle] = useState('');
   const [description, updateDescription] = useState('');
   const [date, updateDate] = useState(`${year}-${month}-${day}`);
-  const [closes, updatedCloses] = useState(`${year}-${month}-${day}`);
+  const [closes, updateCloses] = useState(`${year}-${month}-${day}`);
   const [minGroup, updateMinGroup] = useState(4);
   const [maxGroup, updateMaxGroup] = useState(4);
   const [errors, setErrors] = useState([]);
@@ -34,25 +35,8 @@ export default function NewEventModal () {
       Host: user,
       AttendingUsers: []
     };
-
     dispatch(HardSetList([event]));
-
-    const searchBar = document.querySelector('input.searchbar-input');
-    const searchContainer = document.querySelector('div.searchbar-container');
-    searchBar.style.backgroundColor = 'darkgrey';
-    searchContainer.style.boxShadow = '0px 0px 30px white';
-    let cycles = 0;
-    const attentionInterval = setInterval(() => {
-      (cycles === 2 && clearInterval(attentionInterval)) ?? searchBar.focus();
-      if (cycles % 2) {
-        searchBar.style.backgroundColor = 'darkgrey';
-        searchContainer.style.boxShadow = '0px 0px 20px white';
-      } else {
-        searchBar.style.backgroundColor = 'white';
-        searchContainer.style.boxShadow = 'none';
-      }
-      cycles++;
-    }, 200);
+    FlashSearchbar();
   };
 
   const onSubmit = e => {
@@ -137,7 +121,7 @@ export default function NewEventModal () {
               className='new-event closes'
               type='date'
               value={closes}
-              onChange={({ target: { value } }) => updatedCloses(value)}
+              onChange={({ target: { value } }) => updateCloses(value)}
               required
             />
           </div>
